Lower buffer size, lower google parallelism

+++ src/main/java/fogetti/phish/storm/integration/PhishTopologyRemoteRunner.java
@@ -14,11 +14,11 @@ public class PhishTopologyRemoteRunner {
 		config.setNumWorkers(17);
 		config.setMessageTimeoutSecs(20);
 		config.put(Config.TOPOLOGY_EXECUTOR_RECEIVE_BUFFER_SIZE,
-	               new Integer(131072));
+	               new Integer(8192));
 	    config.put(Config.TOPOLOGY_EXECUTOR_SEND_BUFFER_SIZE,
-	               new Integer(131072));
+	               new Integer(8192));
 	    config.put(Config.TOPOLOGY_TRANSFER_BUFFER_SIZE,
-	               new Integer(131072));
+	               new Integer(8192));
 	    config.registerMetricsConsumer(LoggingMetricsConsumer.class, 1);
 		
 		StormSubmitter.submitTopology("phish-storm", config, topology);
